# Aulas/aula 06/doc/openapi.yaml
openapi: 3.0.3
info:
  title: Projeto de Controle de Filmes - SENAI Jandira
  description: |-
   API destinada para controle de filmes
  termsOfService: http://swagger.io/terms/
  contact:
    email: kayquebrennoferreira@gmail.com
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
...
servers:
  - url: http://
paths:
  /v1/senai/locadora/filme:
    post:
      tags:
        - Filme
      summary: Adiciona um novo Filme
      description: Adiciona um novo Filme
      requestBody:
        description: JSON contendo os dados do novo filme a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Filme'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Filme
      summary: Lista todos os Filmes
      description: Listagem de todos os filmes registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/FilmeGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
  /v1/senai/locadora/genero:
    post:
      tags:
        - Gênero
      summary: Adiciona um novo Genero
      description: Adiciona um novo Genero
      requestBody:
        description: JSON contendo os dados do novo genero a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Genero'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Gênero
      summary: Lista todos os Generos
      description: Listagem de todos os generos registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/GeneroGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
  /v1/senai/locadora/classificacao:
    post:
      tags:
        - Classificação
      summary: Adiciona uma nova Classificacao
      description: Adiciona uma nova Classificacao
      requestBody:
        description: JSON contendo os dados da noca Classificacao a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Classificacao'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Classificação
      summary: Lista todas as Classificacao
      description: Listagem de todas as Classificacao registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ClassificacaoGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
  /v1/senai/locadora/sexo:
    post:
      tags:
        - Sexo
      summary: Adiciona um novo Sexo
      description: Adiciona um novo Sexo
      requestBody:
        description: JSON contendo os dados do Sexo a ser inserido no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Sexo'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Sexo
      summary: Lista todos os Sexos
      description: Listagem de todos os Sexos registrados no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SexoGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
  /v1/senai/locadora/nacionalidade:
    post:
      tags:
        - Nacionalidade
      summary: Adiciona uma nova Nacionalidade
      description: Adiciona uma nova Nacionalidade
      requestBody:
        description: JSON contendo os dados da Nacionalidade a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Nacionalidade'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Nacionalidade
      summary: Lista todas as Nacionalidades
      description: Listagem de todas as Nacionalidades registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/NacionalidadeGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
  /v1/senai/locadora/atividade:
    post:
      tags:
        - Atividade
      summary: Adiciona uma nova Atividade
      description: Adiciona uma nova Atividade
      requestBody:
        description: JSON contendo os dados da Atividade a ser inserida no sistema
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Atividade'
        required: true
      responses:
        '201':
          description: Success Created Item
        '400':
          description: Error Bad Request
        '415':
          description: Error Content Type
        '500':
          description: Error Internal Server (Controller | Model)
    get:
      tags:
        - Atividade
      summary: Lista todas as Atividades
      description: Listagem de todas as Atividades registradas no sistema
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AtividadeGet'  
        '404':
          description: Error Not Found
        '500':
          description: Error Internal Server (Controller | Model)
...
components:
  schemas:
...
    Filme:
      required:
        - nome
        - sinopse
        - capa
        - data_lancamento
        - duracao
      type: object
      properties:
        nome:
          type: string
          description: O número máximo de caracteres é de 80
        sinopse:
          type: string
          description: O número máximo de caracteres é ilimitado
        capa:
          type: string
          description: O número máximo de caracteres é 255
        data_lancamento:
          type: string
          format: date
          example: "2026-01-01"
        duracao:
          type: string
          example: "01:30:00"
          pattern: '^(?:[01]\d|2[0-3]):[0-5]\d$'
        valor:
          type: number
          format: double
          example: "99.99"
          description: O número máximo de caracteres é 999 e duas casas decimais (.99)
        avaliacao:
          type: number
          format: double
          example: "5.0"
          description: O número máximo de caracteres é 5 e uma casa decimal (.9), com o máximo de 5.0
    Genero:
      required:
        - genero
      type: object
      properties:
        genero:
          type: string
          description: O número máximo de caracteres é de 30
    Classificacao:
      required:
        - classificacao
      type: object
      properties:
        classificacao:
          type: string
          description: O número máximo de caracteres é de 6
    Sexo:
      required:
        - sexo
        - sigla
      type: object
      properties:
        sexo:
          type: string
          description: O número máximo de caracteres é de 20
        sigla:
          type: string
          description: O número máximo de caracteres é de 3
    Nacionalidade:
      required:
        - nacionalidade
        - sigla
      type: object
      properties:
        nacionalidade:
          type: string
          description: O número máximo de caracteres é de 90
        sigla:
          type: string
          description: O número máximo de caracteres é de 4
    Atividade:
      required:
        - atividade
      type: object
      properties:
        atividade:
          type: string
          description: O número máximo de caracteres é de 40
...
    FilmeGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nome:
          type: string
          description: O número máximo de caracteres é de 80
          example: Sexta Feira 13
        sinopse:
          type: string
          description: O número máximo de caracteres é ilimitado
        capa:
          type: string
          description: O número máximo de caracteres é 255
          example: url da foto
        data_lancamento:
          type: string
          format: date
          example: 2026-01-01
        duracao:
          type: string
          example: "01:30:00"
          pattern: '^(?:[01]\d|2[0-3]):[0-5]\d$'
        valor:
          type: number
          format: double
          example: 99.99
          description: O número máximo de caracteres é 999 e duas casas decimais (.99)
        avaliacao:
          type: number
          format: double
          example: 5.0
          description: O número máximo de caracteres é 5 e uma casa decimal (.9), com o máximo de 5.0
    GeneroGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        genero:
          type: string
          description: O número máximo de caracteres é de 30
          example: Terror
    ClassificacaoGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        classificacao:
          type: string
          description: O número máximo de caracteres é de 6
          example: +18
    SexoGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        sexo:
          type: string
          description: O número máximo de caracteres é de 20
          example: Masculino
        sigla:
          type: string
          description: O número máximo de caracteres é de 3
          example: M
    NacionalidadeGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        nacionalidade:
          type: string
          description: O número máximo de caracteres é de 90
          example: Brasileiro
        sigla:
          type: string
          description: O número máximo de caracteres é de 4
          example: BR
    AtividadeGet:
      type: object
      properties:
        id:
          type: integer
          example: 1
        atividade:
          type: string
          description: O número máximo de caracteres é de 40
          example: Duble
...